made compatible with the frontend part of the project, the database changed to MySQL

+++ src/main/java/com/muratk/insurancecampaignproject/dataAccess/CampaignRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CampaignRepository extends JpaRepository<Campaign,Integer> {
+public interface CampaignRepository extends JpaRepository<Campaign,Long> {
 
     List<Campaign> findByCampaignCategoryAndCampaignsAdvertTitleAndCampaignsDetails(String category, String title, String details );
